+++ core/src/main/java/dev/failsafe/event/ExecutionAttemptedEvent.java
@@ -41,14 +41,6 @@ public class ExecutionAttemptedEvent<R> extends ExecutionEvent {
   }
 
   /**
-   * @deprecated Use {@link #getLastException()} instead
-   */
-  @Deprecated
-  public Throwable getLastFailure() {
-    return exception;
-  }
-
-  /**
    * Returns the result that preceded the event, else {@code null} if there was none.
    */
   public R getLastResult() {