+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### API Changes
 
 - `ExecutionContext.getStartTime` now returns an `Instant` rather than a `Duration` object.
+- `getFailure`, `getLastFailure`, `recordFailure` and similar methods for recording Exceptions, which were previously deprecated, were removed. Use `getException`, `getLastException`, `recordException`, etc. instead.
 
 # 3.2.4